继续优化--后台 遇到问题，string.trim 去不掉空格？？疑问

+++ BlazUICommunity/BlazUICommunity.Api/Controllers/Admin/BannerController.cs
@@ -7,12 +7,14 @@ using AutoMapper;
 using Blazui.Community.Api.Extensions;
 using Blazui.Community.Api.Service;
 using Blazui.Community.DTO;
+using Blazui.Community.DTO.Admin;
 using Blazui.Community.Model.Models;
 using Blazui.Community.Repository;
 using Blazui.Community.Request;
 using Blazui.Community.Utility;
 using Blazui.Community.Utility.Filter;
 using Blazui.Community.Utility.Response;
+using Marvin.Cache.Headers;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Swashbuckle.AspNetCore.Annotations;
@@ -26,6 +28,7 @@ namespace Blazui.Community.Api.Controllers
     [HiddenApi]
     [ApiController]
     [SwaggerTag(description: "banner")]
+    [HttpCacheExpiration(MaxAge = 100)]
     public class BannerController : ControllerBase
     {
 
@@ -58,27 +61,36 @@ namespace Blazui.Community.Api.Controllers
         }
 
 
-
-        /// <summary>
-        /// 逻辑删除
-        /// </summary>
-        /// <returns></returns>
-        [HttpDelete("Delete/{Id}")]
-        public IActionResult Delete([FromRoute] string Id)
+        private async Task<IActionResult> DeleteOrResume(string Id, int Status)
         {
-            //_BZBannerRepository.LogicDelete(Id, oprationId);
-            //return Ok();
-
-            var version = _BZBannerRepository.Find(Id);
-            if (version != null && !string.IsNullOrWhiteSpace(version.Id))
-            {
-                version.Status = version.Status == -1 ? 0 : -1;
-            }
-            _BZBannerRepository.Update(version);
+            var banner = await _BZBannerRepository.FindAsync(Id);
+            if (banner is null)
+                return BadRequest();
+            if (banner.Status == Status)
+                return Ok();
+            banner.Status = Status;
+            _BZBannerRepository.Update(banner);
             _cacheService.Remove(nameof(BzBannerModel));
             return Ok();
         }
-
+        /// <summary>
+        /// 删除
+        /// </summary>
+        /// <returns></returns>
+        [HttpPatch("Delete/{Id}")]
+        public async Task<IActionResult> Delete([FromRoute] string Id)
+        {
+            return await DeleteOrResume(Id, -1);
+        }
+        /// <summary>
+        /// 恢复
+        /// </summary>
+        /// <returns></returns>
+        [HttpPatch("Resume/{Id}")]
+        public async Task<IActionResult> Resume([FromRoute] string Id)
+        {
+            return await DeleteOrResume(Id, 0);
+        }
         /// <summary>
         /// 更新
         /// </summary>
@@ -93,56 +105,19 @@ namespace Blazui.Community.Api.Controllers
             _cacheService.Remove(nameof(BzBannerModel));
             return Ok();
         }
+     
 
-        /// <summary>
-        /// 根据Id查询
-        /// </summary>
-        /// <returns></returns>
-        [HttpGet("Query/{Id}")]
-        public async Task<IActionResult> Query([FromRoute] string Id)
-        {
-            var res = await _BZBannerRepository.FindAsync(Id);
-            if (res is null)
-                return NoContent();
-            return Ok(res);
-        }
         /// <summary>
         /// 根据条件分页查询
         /// </summary>
         /// <returns></returns>
-        [HttpPost("Query")]
-        public async Task<IActionResult> Query([FromBody] BannerRequestCondition Request = null)
+        [HttpGet("Query")]
+        public async Task<ActionResult<IPagedList<BannerDisplayDto>>> Query([FromQuery] BannerReuestCondition reuestCondition)
         {
-            IPagedList<BzBannerModel> pagedList = null;
-            var query = Request.CreateQueryExpression<BzBannerModel, BannerRequestCondition>();
-            pagedList = await _BZBannerRepository.GetPagedListAsync(query, o => o.OrderBy(p => p.Id), null, Request.PageIndex - 1, Request.PageSize);
-            var pagedatas = pagedList.ConvertToPageData<BzBannerModel, BzBannerDto>();
-            pagedatas.Items = _mapper.Map<IList<BzBannerDto>>(pagedList.Items);
+            var pagedList = await _BZBannerRepository.GetPagedListAsync(p => true, o => o.OrderBy(p => p.Id), null, reuestCondition.PageIndex - 1, reuestCondition.PageSize);
+            var pagedatas = pagedList.From(result => _mapper.Map<IList<BannerDisplayDto>>(result));
             return Ok(pagedatas);
         }
 
-        /// <summary>
-        /// 根据条件分页查询
-        /// </summary>
-        /// <returns></returns>
-        [HttpGet("GetPageList/{pageSize}/{pageIndex}")]
-        public async Task<IActionResult> Query(int pageSize, int pageIndex)
-        {
-
-            var pagedList = await _BZBannerRepository.GetPagedListAsync(p =>true, o => o.OrderBy(p => p.Id), null, pageIndex - 1, pageSize);
-            var pagedatas = pagedList.ConvertToPageData<BzBannerModel, BzBannerDto>();
-            pagedatas.Items = _mapper.Map<IList<BzBannerDto>>(pagedList.Items);
-            return Ok(pagedatas);
-        }
-        /// <summary>
-        /// 根据条件分页查询
-        /// </summary>
-        /// <returns></returns>
-        [HttpGet("QueryAll")]
-        public async Task<IActionResult> QueryAll()
-        {
-            var banners = await _cacheService.Banners(p => p.Show && p.Status == 0);
-            return Ok(banners);
-        }
     }
 }
